% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{bp}
\alias{bp}
\title{Simulate multi-type Markov branching process}
\usage{
bp(e_mat, r_vec, p_vec, z0_vec, times)
}
\arguments{
\item{e_mat}{The matrix of birth events that can occur in the brnaching process. Dimensions \code{nevents} x \code{ntypes}}

\item{r_vec}{A vector containing the rate at which each of the birth events in \code{e_mat} occurs. Dimensions \code{nevents} x 1}

\item{p_vec}{A vector containing the parent type for each of the birth events in \code{e_mat}. Dimensions \code{nevents} x 1}

\item{z0_vec}{The initial population vector at time 0. Dimensions \code{ntypes} x 1}

\item{times}{The timepoints at which to record the population}
}
\value{
A matrix with the population vectors at each timepoint. Dimensions \code{ntimes} x \code{ntypes}
}
\description{
Simulate multi-type Markov branching process
}
