% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bp_model.R
\name{bp_model_simple_birth_death}
\alias{bp_model_simple_birth_death}
\title{Convenience constructor for building a simple 1-type birth-death process}
\usage{
bp_model_simple_birth_death(func_deps, nparam, ndep)
}
\arguments{
\item{func_deps}{A vector of expressions encoding how each rate depends on the parameters of the model and a set of dependent variables.}

\item{nparam}{The number of parameters in the model}

\item{ndep}{The number of dependent variables which influence the rates of the model}
}
\value{
A new branching process model object if the parameters passed are valid. Otherwise an exception is thrown.
}
\description{
Convenience constructor for building a simple 1-type birth-death process
}
