% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{bpsims}
\alias{bpsims}
\title{Simulate multi-type Markov branching process where process rates vary as a function of dependent variables}
\usage{
bpsims(model, theta, z0_vec, times, reps, c_mat = NA)
}
\arguments{
\item{model}{An object of type \code{bpmodel} representing the model to be simulated}

\item{theta}{A vector of dependent variables to calculate the rates of each birth venet in \code{model}. Dimensions \code{ndep_levels} x \code{ndep}}

\item{z0_vec}{The initial population vector at time 0. Dimensions \code{ntypes} x 1}

\item{times}{The timepoints at which to record the population}

\item{reps}{a vector containing the number of times to replicate each dependent variables condidion. Dimensions \code{ndep_levels} x 1}
}
\value{
A matrix with the population vectors at each timepoint. Dimensions \code{ntimes} x \code{ntypes}
}
\description{
Simulate multi-type Markov branching process where process rates vary as a function of dependent variables
}
