% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_utils.R
\name{check_valid}
\alias{check_valid}
\title{takes simple mathematical R expressions and determines whether they can be parsed into stan code}
\usage{
check_valid(exprn, max_params, max_dep)
}
\arguments{
\item{exprn}{the expression to parse}

\item{max_params}{the number of parameters in the model, which determines the maximum parameter index that can be referenced in the expression}

\item{max_dep}{the number of dependent variables in the model, which determines the amximum dependent variables index that can be referenced in the model}
}
\value{
Nothing if the expression is valid. Otherwise an exception is thrown.
}
\description{
takes simple mathematical R expressions and determines whether they can be parsed into stan code
}
