% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_utils.R
\name{exp_to_stan}
\alias{exp_to_stan}
\title{takes simple mathematical R expressions and turns them into a specific type of Stan code to fill in the template}
\usage{
exp_to_stan(exprn, max_params, max_dep)
}
\arguments{
\item{exprn}{the expression to parse}

\item{max_params}{the number of parameters in the model, which determines the maximum parameter index that can be referenced in the expression}

\item{max_dep}{the number of dependent variables in the model, which determines the amximum dependent variables index that can be referenced in the model}
}
\value{
A string with the Stan code derived from the R expressions
}
\description{
takes simple mathematical R expressions and turns them into a specific type of Stan code to fill in the template
}
