% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bp_model.R
\name{new_bp_model}
\alias{new_bp_model}
\title{model class that can be used for inference and simulation
constructor for object of class \code{bp_model}}
\usage{
new_bp_model(e_mat, p_vec, func_deps, nparam, ndep)
}
\arguments{
\item{e_mat}{The matrix of birth events that can occur in the brnaching process. Dimensions \code{nevents} x \code{ntypes}}

\item{p_vec}{A vector containing the parent type for each of the birth events in \code{e_mat}. Dimensions \code{nevents} x 1}

\item{func_deps}{A vector of expressions encoding how each rate depends on the parameters of the model and a set of dependent variables.}

\item{nparam}{The number of parameters in the model}

\item{ndep}{The number of dependent variables which influence the rates of the model}
}
\value{
A new branching process model object
}
\description{
model class that can be used for inference and simulation
constructor for object of class \code{bp_model}
}
