% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_utils.R
\name{parse_prior}
\alias{parse_prior}
\title{parses prior object into a prior statement that can be inserted into a Stan model (e.g. Theta4 ~ normal(0,1))}
\usage{
parse_prior(prior, k)
}
\arguments{
\item{prior}{the \code{prior_dist} object to be parsed}

\item{k}{the number parameter corresponding to the given prior.}
}
\value{
2 strings: A parameter declaration the goes in the \code{parameters} block of the Stan model, 
and a prior probability statement that goes in the \code{model} block of the Stan model
}
\description{
parses prior object into a prior statement that can be inserted into a Stan model (e.g. Theta4 ~ normal(0,1))
}
