% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_utils.R
\name{prior_dist}
\alias{prior_dist}
\title{User-facing constructor for new \code{prior_dist} objects}
\usage{
prior_dist(name, params, bounds = NA)
}
\arguments{
\item{name}{a string with the name of the distribution (e.g. 'normal'). Must match a distribution in \code{ALLOWED_DISTRIBUTIONS} list}

\item{params}{parameters for the prior distribution. Should match the number and constrains imposed by the distribution name.}

\item{bounds}{a lower and upper bound for the parameter. This restricts the range in which sampling occurs. If left as NA, no bounds are imposed}
}
\value{
a new \code{prior_dist} object in the arguments specify a valid distribution. Throws an error otherwise.
}
\description{
User-facing constructor for new \code{prior_dist} objects
}
