% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_utils.R
\name{stan_data_from_simulation}
\alias{stan_data_from_simulation}
\title{Helper function for piping simulation output into inference engine}
\usage{
stan_data_from_simulation(sim_data, model)
}
\arguments{
\item{sim_data}{simulation data as returned from \code{bpsims}}

\item{model}{the \code{bpmodel} object the produced the data}
}
\value{
A data list to pass to the Stan sampling function
}
\description{
Helper function for piping simulation output into inference engine
}
