% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_utils.R
\name{uniform_initialize}
\alias{uniform_initialize}
\title{create Stan initialization list by selecting initial values uniformly at ranom.}
\usage{
uniform_initialize(ranges, nchains)
}
\arguments{
\item{nchains}{number of Markov chains for which to sample initial values}

\item{range}{nparams x 2 matrix, where nparams is the number ofparameters. Each row has a lower and upper bound for initialization}
}
\value{
a stan initialization list
}
\description{
create Stan initialization list by selecting initial values uniformly at ranom.
}
