% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bp_model.R
\name{validate_bp_model}
\alias{validate_bp_model}
\title{function to determine whether an object is a valid instance of class \code{bpmodel}}
\usage{
validate_bp_model(bpm)
}
\arguments{
\item{bpm}{The branching process model object}
}
\value{
The model passed into the function if it is vaild. Throws an exception otherwise
}
\description{
function to determine whether an object is a valid instance of class \code{bpmodel}
}
