% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_utils.R
\name{validate_prior_dist}
\alias{validate_prior_dist}
\title{function to make sure a \code{prior_dist} object actually encodes a vaild prior distribution}
\usage{
validate_prior_dist(prior)
}
\arguments{
\item{prior, }{the \code{prior_dist} object being validated}
}
\value{
the \code{prior_dist} if the prior is valid. Will throw an error otherwise.
}
\description{
function to make sure a \code{prior_dist} object actually encodes a vaild prior distribution
}
